added setters for exported variable, slide userSearchScreen works and refactored a bit of code
