+++ backend/html/index.html
@@ -238,9 +238,10 @@
     <div class="planetPanel">
       <img src="{% static 'game/assets/scoreUI.png' %}" class="blueShadowfilter">
       <img src="{% static 'game/assets/scoreUI.png' %}" class="blueShadowfilter">
-      <div class="background blueShadowfilter" style="flex-direction: column;">
+      <div class="background blueShadowfilter" style="flex-direction: column; align-items: flex-start;">
         <div id="arenaTitleText" style="font-size: 50px; font-family: 'space'; color: white; margin-top: -15px;"> Profile </div>
-          <div class="userPage">
+        <div class="userPagesContainer">
+          <div id="hostUserPage" class="userPage">
             <div class="rightColumn" style="width: 75%;">
               <div class="aboveStatsBlock">
                 <div class="leftBlock">
@@ -255,19 +256,19 @@
                         <img src="{% static 'html/assets/icons/pencil.png' %}">
                       </div>
                     </div>
-                      <div class="winLoseTexts">
+                    <div class="winLoseTexts">
                         <div style="font-family: 'Space'; font-size: 20px; color: white"> Parties : 13</div> 
                         <div style="font-family: 'Space'; font-size: 20px; color: white"> Victoires : 5</div> 
                         <div style="font-family: 'Space'; font-size: 20px; color: white"> Defaites : 8</div>
                         <div style="font-family: 'Space'; font-size: 20px; color: white"> Buts : 33</div>
+                      </div>
                     </div>
-                  </div>
                   <div class="statsModesBlock">
                    <p>statistics</p> 
-                    <div class="statsModes">
-                      <div class="statButton" style="border-top-right-radius: 0px; border-bottom-right-radius: 0px; background-color: #5d75ff47;">Stat1</div>
-                      <div class="statButton" style="border-radius: 0px;">Stat2</div>
-                      <div class="statButton" style="border-top-left-radius: 0px; border-bottom-left-radius: 0px;">Stat3</div>
+                   <div class="statsModes">
+                     <div class="statButton" style="border-top-right-radius: 0px; border-bottom-right-radius: 0px; background-color: #5d75ff47;">Stat1</div>
+                     <div class="statButton" style="border-radius: 0px;">Stat2</div>
+                     <div class="statButton" style="border-top-left-radius: 0px; border-bottom-left-radius: 0px;">Stat3</div>
                     </div>
                   </div>
                 </div>
@@ -288,20 +289,70 @@
         </div>
       </div>
     </div>
+    <div id="exploreUserPage" class="exploreUserPage">
+      <div class="rightColumn" style="width: 75%;">
+        <div class="aboveStatsBlock">
+          <div class="leftBlock">
+            <div class="userInfoBlock">
+              <div class="nameImgBlock" id="profile_info">
+                <div class="imgFrame"><img id="profile_pic" src="" alt="Photo de profil"></div> 
+                <div class="mail-pseudo-info">
+                  <div id="username" style="color: rgb(255, 255, 255); "></div> 
+                  <div id="bio" style="font-size: 10px; color: rgb(189, 209, 255); overflow-wrap: break-word;"></div> 
+                </div>
+                <div class="pencilImg">
+                  <img src="{% static 'html/assets/icons/pencil.png' %}">
+                </div>
+              </div>
+              <div class="winLoseTexts">
+                  <div style="font-family: 'Space'; font-size: 20px; color: white"> Parties : 13</div> 
+                  <div style="font-family: 'Space'; font-size: 20px; color: white"> Victoires : 5</div> 
+                  <div style="font-family: 'Space'; font-size: 20px; color: white"> Defaites : 8</div>
+                  <div style="font-family: 'Space'; font-size: 20px; color: white"> Buts : 33</div>
+                </div>
+              </div>
+            <div class="statsModesBlock">
+              <p>statistics</p> 
+              <div class="statsModes">
+                <div class="statButton" style="border-top-right-radius: 0px; border-bottom-right-radius: 0px; background-color: #5d75ff47;">Stat1</div>
+                <div class="statButton" style="border-radius: 0px;">Stat2</div>
+                <div class="statButton" style="border-top-left-radius: 0px; border-bottom-left-radius: 0px;">Stat3</div>
+              </div>
+            </div>
+          </div>
+          <div class="rightBlock">
+          <div id="historyText" style="font-size: 20px; font-family: 'space'; color: white; margin-bottom: 4px">History</div>
+            <div class="history">
+            </div>
+          </div>
+        </div>
+        <div class="statsBlock"></div>
+      </div>
+<div class="leftColumn" style="width: 22%; height: 100%; justify-content: space-between;">
+  <div id="FriendListText" style="font-size: 20px; font-family: 'space'; color: white"> Friends List </div>
+  <div class="friendSearchBar">
+    <input id="searchInput" class="text-input2" style="font-size: 15px;" spellcheck="false" placeholder="Look for users . . ."></input>
+  </div>
+  <div id="userlistUserPage" class="userlistBackground" style="height:85%; width: 100%; border-radius: 10px;">
   </div>
 </div>
+</div>  
+</div>
+</div>
+</div>  
+</div>
 
-    <!-- ESCAPE PAGE -->
-    <div class="coloredPanel"></div>
-    <div class="structure">
-        <div class="pauseTitle">
-          <img src="{% static 'game/assets/scoreUI.png' %}">
-          PAUSE
-          <img src="{% static 'game/assets/scoreUI.png' %}">
-        </div>
-        <div class="escapeBG">
-        <div class="settings">
-          <div class="languageColumn" id="graphicsText" style="width: 50%; height: 40%;">
+<!-- ESCAPE PAGE -->
+<div class="coloredPanel"></div>
+<div class="structure">
+  <div class="pauseTitle">
+    <img src="{% static 'game/assets/scoreUI.png' %}">
+    PAUSE
+    <img src="{% static 'game/assets/scoreUI.png' %}">
+  </div>
+  <div class="escapeBG">
+    <div class="settings">
+      <div class="languageColumn" id="graphicsText" style="width: 50%; height: 40%;">
             <div class="languageSettings">
             <div class="graphicsIcon" id="graphicsIcon1">Simple</div>
             <div class="graphicsIcon redGlow" id="graphicsIcon2">Classic</div>
