+++ backend/static/html/css/planetPage.css
@@ -18,7 +18,7 @@
 	position: absolute;
 	transform: translate(-50%, -50%) rotate(-90deg);
 	top: 50%;
-	left: 50%;;
+	left: 50%;
 	display: flex;
 	justify-content: center;
 	align-items: center;
