+++ backend/static/html/css/userPage.css
@@ -1,4 +1,24 @@
+.userPagesContainer  {
+    border: 2px solid white;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    flex-direction: row;
+    width: 200%;
+    height: 90%;
+}
+
 .userPage {
+    /* border: 2px solid white; */
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    flex-direction: row;
+    height: 90%;
+    width: 100%;
+}
+
+.exploreUserPage {
     /* border: 2px solid white; */
     display: flex;
     align-items: center;
@@ -6,6 +26,15 @@
     flex-direction: row;
     height: 90%;
     width: 100%;
+}
+
+@keyframes slideUserPage {
+    from {
+        transform: translateX(0); /* Start position */
+    }
+    to {
+        transform: translateX(-50.8%); /* Slide to the right */
+    }
 }
 
 .statsBlock {
@@ -270,12 +299,19 @@
     justify-content: center;
     height: 100%;
     width: 25%;
+    background-color: #5d75ff47;
+    border: 2px solid #3777ff;
+    border-radius: 5px;
 }
 
 .loupeImg img {
     height: auto;
     width: 50px;
     transform: rotate(-30deg);
+}
+
+.loupeImg:hover {
+    background-color: #000f61b2;
 }
 
 .imgFrame img {
